+++ App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet} from 'react-native';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 
 import SplashScreen from './screens/SplashScreen';
