+++ screens/HomeScreen.js
@@ -1,27 +1,18 @@
 import React, { Component } from "react";
 import { View, Text, StyleSheet, Button } from "react-native";
-import firebase from 'firebase';
+import { addUser, userSignOut } from '../utils/authentication/user.js';
 
 class HomeScreen extends Component {
     state = { user: {} };
     componentDidMount() {
-        this.addUser();
-    }
-
-    addUser() {
-        firebase.auth().onAuthStateChanged((user) => {
-            if (user != null) {
-                this.setState({ user: user });
-            }
-        })
+        addUser.bind(this);
     }
 
     render() {
         return (
             <View style={styles.container}>
                 <Text>HomeScreen</Text>
-                <Button title="Sign out" onPress={() => {
-                    firebase.auth().signOut();}} />
+                <Button title="Sign out" onPress={userSignOut} />
             </View>
         )
     }
