+++ utils/authentication/login.js
@@ -1,0 +1,8 @@
+export const onSuccessfulLogin = (props) => {
+    props.navigation.navigate('HomeScreen');
+}
+
+
+export const onLoginFailure = (errorMessage) => {
+    this.setState({ error: errorMessage, loading: false });
+}
