+++ utils/authentication/user.js
@@ -1,0 +1,23 @@
+import firebase from 'firebase';
+
+export const checkUserLoggedIn = (props) => {
+    firebase.auth().onAuthStateChanged(user => {
+        if (user) {
+            props.navigation.navigate('HomeScreen');
+        } else {
+            props.navigation.navigate('LoginScreen');
+        }
+    })
+}
+
+export const addUser = () => {
+    firebase.auth().onAuthStateChanged((user) => {
+        if (user != null) {
+            this.setState({ user: user });
+        }
+    })
+}
+
+export const userSignOut = () => {
+    firebase.auth().signOut();
+}
